Get user's report for multiple repositories (#26)
Added ability to list more than one repository for user metrics. The
system now accepts a new input of `owner/repo` style, and if more repos
are added they have to be separated by a comma.

Resolves #21

If the repo metrics are invoked, the system will only use the first
repo, ignoring any other available repos.

The repo information is still being cached in `.cache` and are combined
on runtime to generate a bigger sample.